% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hoxompkgs}
\alias{hoxompkgs}
\title{HOXOM family to development and maintain packages.}
\format{A data frame with 22 rows and 2 variables:
\itemize{
  \item{full_name: GitHub repository name}
  \item{description: Package description}
}}
\usage{
hoxompkgs
}
\description{
HOXOM family to development and maintain packages.
}
\keyword{datasets}

