% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htv2html.R
\name{htv2html}
\alias{htv2html}
\title{Generare Hoxom task view for render as html.}
\usage{
htv2html(file = NULL)
}
\arguments{
\item{file}{character}
}
\description{
Wrapper function of ctv::ctv2html()
}
\examples{
\dontrun{
htv2html()
httr::BROWSE("Hoxom.html")
}
}

