% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_hoxom_tv.R
\name{install_hoxom_tv}
\alias{install_hoxom_tv}
\title{Install package from hoxom-task view list}
\usage{
install_hoxom_tv()
}
\description{
Install package from hoxom-task view list
}
\examples{
\dontrun{
install_hoxom_tv()
}
}

