% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_Hoxom_tv.R
\name{mk_Hoxom_tv}
\alias{mk_Hoxom_tv}
\title{Making Hoxom task view}
\usage{
mk_Hoxom_tv()
}
\description{
Making Hoxom task view
}
\examples{
\dontrun{
mk_Hoxom_tv()

mk_Hoxom_tv() \%>\% saveXML(file = "Hoxom.ctv")
}
}

